resolved NOTWORKING errors
compilation was failing for Eigen::SparseMatrix + Eigen::SparseMatrix arguments since the TENSOR_SPMAT_OP macro yielded two C++ functions that could handle these as arguments (so, compiler was faced with an ambiguous function call).  But, the TENSOR_SPMAT_OP macro was designed for use with a tensor or tensor expression as one of the inputs (and an Eigen::SparseMatrix as the other input).

Resolved issue by using SFINAE to update the macro to be more specific in when it would apply